The documentation is pretty straightforward in saying what goes where: https://community.compuwareapm.com/community/display/DOCDT60/Tagging+ADK+for+.NET
In fact, the documented way to go about this is to initialize both the DynaTraceADKFactory and Tagging instances at process startup.  I think because you’re working with a multi-threaded application it is likely better to do this per-thread, so initializing at the start of this class makes sense.  The rest includes necessary checks and balances, and if working properly should probably not be modified.